ai-agent/requirements.txt
livekit-agents>=0.8.0
livekit-plugins-groq
livekit-plugins-elevenlabs
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-cartesia
livekit-plugins-silero
python-dotenv
requests
pydub
openai
groq
pydantic